Update to version 2.4.1 of MFizz.
Also introduce a small Python script to automatically generate the .sty
file from the .css file inside the [font-mfizz
repository](https://github.com/fizzed/font-mfizz).

This should fix #1.

In the long run, the .ttf file should also be taken from the submodule.

A CTAN release of this updated version is on the way.